[TASK] Adjust domain model
The domain model of Media extends the model of FAL with its own fields.
It needs to be adjusted to the final version of FAL. Properties can be
accessed through the getProperty method.

Additionally, all fields of Media must be respected in the model.

With this change, all fields except categories and variants (that
are object storages) are respected and therefore available in Fluid
with the usual syntax.

Change-Id: I908ab9b8aaffaefd64a69f70073f7886387378c8
Related: #41636
Release: 1.0