[!!!][BUGFIX] Use xmlns attribute for Media namespace
Beware this patch only apply on TYPO3 6.1 and break the backward compatibiliy
with TYPO3 6.0 for the BE module. Fluid has improved the way namespace can
be declared by making the xmlns attibute clever. Taking advantage of this
makes the traditional way [1] of declaring namespace redundant
and, as a result, raises a PHP exception.

[1] {namespace m=TYPO3\CMS\Media\ViewHelpers}

Change-Id: I2da24fdadb949efaed661fc4ec2d3a4bc94841ff
Fix: #47116
Reviewed-on: https://review.typo3.org/20022
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot