[TASK] Make FineUploader components translatable
The FineUploader components (drag and drop on the file TCEFORM and the
Media module, upload buttons in the file TCEFORM and Media module) are
currently not translatable and using the default labels of FineUploader.
This is changed.

At the same time the CSS of the FineUploader is adapted to look better.
The drag and drop upload in the TCEFORM now looks like similar to the
one in the File List, while the drag and drop upload in the Media module
was made usable at all.

Resolves: #59565
Releases: 3.0
Change-Id: Ia00fd6e408bfb818a2f102392ae6f4084eb82a7c
Reviewed-on: https://review.typo3.org/30719
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot