[FEATURE] Refactor find by Usage to be more efficient
To be efficient on large data set of files,
a caching column "number_of_references" had
to be introduced in "sys_file".
The column is updated by Hook whenever a record
containing a File reference in saved in TCEMain.

On a fresh install, consider initializing
the values by the CLI command:

./typo3/cli_dispatch.phpsh extbase media:warmUpCache

Change-Id: Id6db413f6890c879dccf548e07f72731777ebc57
Release: 3.0
Resolves: #61141
Reviewed-on: http://review.typo3.org/32313
Reviewed-by: Fabien Udriot <fabien.udriot@ecodev.ch>
Tested-by: Fabien Udriot <fabien.udriot@ecodev.ch>