[TASK] Remove the Asset Domain Model
For CMS 6.1, Media had its own API for querying
the File according to various criteria.

However, after the split of the tables sys_file and
sys_file_metadata in CMS 6.2, the API could not be resurrected.

As a result, this patch remove the last bit
of the Domain Model from the extension - which is not used anyhow.

The Vidi Content Repository could be used instead to retrieve data
in a very flexible way.

Change-Id: I06ce00459f143a78a455aa4a3f9c344b9f5389f3
Releases: 3.0
Resolves: #61346
Reviewed-on: http://review.typo3.org/32579
Reviewed-by: Fabien Udriot <fabien.udriot@ecodev.ch>
Tested-by: Fabien Udriot <fabien.udriot@ecodev.ch>