[WIP/FEATURE] Support TYPO3 Flash-Messages
Currently, if upload fails, it just cumulates the flash error messages and prints them on the next page (which might or might not be related to the upload page). We tried to add output of these messages, but it's probably not a good idea to modify fineuploader. However we failed to find a place where the changes would fit in better.