[BUGFIX] Preserve storage selection menu when editing a record
When inserting a media into RTE, a popup window will open to select the media.
This window shows a selectbox menu for storage selection, but this menu
disappeared if user edited a record and closed it. Instead of the menu a
checkbox "Browse subfolders" was displayed.

By forwarding specific parameters to the edit button `returnUrl` parameter,
we are now able to preserve the storage menu selection, when the user comes
back from the edit page.

Resolves #181
Resolves #183