[BUGFIX] flushProcessedFilesCommand: Don't try flushing files if storage is not publicly available
The flushProcessedFiles command flushes all processed files of all storages.

To find the appropriate folder, the publicUrl of the Storage's processingFolder is looked up.
However, there is no publicUrl for a Storage that is not publicly available. This is usually the
case if you have protected file areas.

In this case, the contents of PATH_SITE are deleted which is no desirable. With this change,
the command skips storages that have no publicly available processingFolder.

Resolves: #209